Improve mod release announcement Discord webhook
